Tausendertrennzeichen bei Gesamtbeträgen

+++ database/migrations/2019_02_04_092027_create_prescribing__suggestions_table.php
@@ -24,7 +24,7 @@ class CreatePrescribingSuggestionsTable extends Migration
             $table->string('title');
             $table->text('description');
             $table->integer('author_id');
-            $table->integer('total_amount');
+            $table->float('total_amount', 8, 2);
             $table->timestamps();
         });
     }
